# Vectorize DotProductImplementation with Vector<T>