### Motivation and Context

Accelerate the DotProductImplementation for floats and doubles using SIMD.

### Description

Employ `Vector<T>` to accelerate DotProductImplementation. If the hardware supports acceleration, a vector's worth of data is processed at a time until there isn't a vector's worth remaining. Then it falls back to the scalar path that's also used if hardware acceleration isn't available.

Note that this removes the manual loop unrolling for lengths 2 and 4.  That would only still be useful on machines that don't have hardware acceleration support, which would be very old machines. The length 2 implementation also had a bug that prevented it from being used at all... it's up-front guard check was doing % 4 instead of % 2, so if you did call DotProduct with a length that was a multiple of 2 but not of 4 such that it fell back to using this implementation, it would immediately throw an exception.  I added a test that would fail with the old implementation.

### Contribution Checklist
<!-- Before submitting this PR, please make sure: -->
- [x] The code builds clean without any errors or warnings
- [x] The PR follows SK Contribution Guidelines (https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
- [x] The code follows the .NET coding conventions (https://learn.microsoft.com/dotnet/csharp/fundamentals/coding-style/coding-conventions) verified with `dotnet format`
- [x] All unit tests pass, and I have added new tests where possible
- [ ] I didn't break anyone :smile:

On my machine:

```C#
private double[] _values1, _values2;

[Params(2, 4, 16, 130, 1_024)]
public int Length { get; set; }

[GlobalSetup]
public void Setup()
{
    _values1 = Enumerable.Range(0, Length).Select(_ => Random.Shared.NextDouble()).ToArray();
    _values2 = Enumerable.Range(0, Length).Select(_ => Random.Shared.NextDouble()).ToArray();
}

[Benchmark]
public double Old() => DotProductImplementation1(_values1, _values2);

[Benchmark]
public double New() => DotProductImplementation2(_values1, _values2);
```

Floats:

| Method | Length |         Mean |
|------- |------- |-------------:|
|    Old |      2 |     6.688 ns |
|    New |      2 |     4.683 ns |
|    Old |      4 |     8.324 ns |
|    New |      4 |     5.594 ns |
|    Old |     16 |    13.444 ns |
|    New |     16 |     7.326 ns |
|    Old |    130 |    64.490 ns |
|    New |    130 |    27.029 ns |
|    Old |   1024 |   432.598 ns |
|    New |   1024 |   192.221 ns |
|    Old |  10240 | 4,717.146 ns |
|    New |  10240 | 1,927.729 ns |

Doubles:

| Method | Length |         Mean |
|------- |------- |-------------:|
|    Old |      2 |     6.832 ns |
|    New |      2 |     4.724 ns |
|    Old |      4 |     7.298 ns |
|    New |      4 |     4.205 ns |
|    Old |     16 |    12.143 ns |
|    New |     16 |     7.881 ns |
|    Old |    130 |    59.685 ns |
|    New |    130 |    34.040 ns |
|    Old |   1024 |   269.562 ns |
|    New |   1024 |   245.630 ns |
|    Old |  10240 | 3,080.644 ns |
|    New |  10240 | 2,816.317 ns |

cc: @tannergooding